File move comment, and start adoption of relative imports
